+++ script/Cargo.toml
@@ -27,7 +27,6 @@ serde = { version = "1", features = ["derive"] }
 serde_json = "1.0.86"
 tokio = { version = "1.2.0", features = ["full"] }
 reqwest = "0.11.20"
-ethers = { version = "2.0.10", features = ["ws"] }
 sha256 = "1.4.0"
 primitive-types = "0.12.1"
 
@@ -48,6 +47,7 @@ clap = "4.4.9"
 futures = "0.3.30"
 async-trait = "0.1.77"
 ed25519-dalek = "2.1.1"
+ethers = "2.0.14"
 
 [build-dependencies]
 sp1-helper = { git = "https://github.com/succinctlabs/sp1.git", branch = "dev" }
